Allow for rotation of lucite and light guide (independently).
This replaces the implementation with <trd> with <trap> which does
not require bottom and top face to be centered on the same z axis.
Now, an angle can be specified to generate a mitered edge, i.e. /_/

The miter angle can be applied to the lucite planes (and the glue,
mylar and whole shebang), and a different miter angle to the light
guide and aluminum reflector etc.

The previous configuration can be obtained for both angles set to 0
(no miter on the lucite cuts).

Unfortunately geant4 restrictions prevent the lucity rotation from
working. Although the replicated volume of lucite planes does fill
the parent volume exactly, geant4 calculates N * shift and compares
this with the bounding box of the parent volume. For these rotated
trapezoids the bounding box is defined by the top and bottom edges
and clearly larger than N * shift. Upstream bug fix in progress.
This bug does currently restrict the miter angle on the lucite to
zero, since otherwise it failes at voxel creation. A local bug fix
is unlikely because G4SmartVoxelHeader is deep inside the geometry
management sector of geant4 and can't be locally inherited and
modified in remoll (the way we could modify G4GDMLParser).

Un-mitered rotations can still be applied with the angle in the parent
file pionDetectorLuciteSector.gdml. That rotates the whole lucite/
lightguide/PMT assembly.