Apparently aux tag DetType is squashed.
This, for a single sensitive detector:
  <auxiliary auxtype="DetType" auxvalue="opticalphoton"/>
  <auxiliary auxtype="DetType" auxvalue="lowenergyneutral"/>
is squashed and only the last one is used.

So, this modified opticalphoton to set lowenergyneutral as well,
since (you guessed it) optical photons are low energy neutrals...

Now we get optical photons at the cathode again, at least on my
version of geant4.